Add option to clean www dir
If the option "clean_html" is set to true in rc.lua lenie will now
delete all files in the www directory that are not part of the blog in
its current state. This will come in handy when a post is removed from
the blog (via git rm) and lenie will automatically remove the
corresponding HTML file.