Added debug print statement
During generation of HTML code, when calling markdown, the command to
call markdown with all arguments is now printed to stdout when "verbose"
is set to true in rc.lua.

This is a rather critical section of lenie and due to observed crashes
with huge amounts of files in that particular area it is essential for
debugging to know what the markdown command looked like right before the
crash.