Add inclusion of footer.md
Just as lenie is looking for a preamble.md in the working dir to create
the blogs header from it, she is now also looking for a footer.md to
create the blogs footer.

Additionally fixed a bug where a div container was opened for the HTML
code for every post, but was never closed.