Add function to get sha1 of currently checked out revision
The sha1 of the currently checked out rev is stored in a file named
"rev" in the src dir. So far it has only been read in one place, but for
the next change it will be read somewhere else and in preparation the
functionality to get this sha is now encapsulated in its own function.