First draft of re-written HTML generation
This is a huge commit, re-writing almost all code to generate the HTML
code. All previous code is still there but commented out, the function
gen_html is kept very simple now, only calling other functions and
passing tables between them. These functions are replacement of all the
functions called in the previous gen_html-function.

This commit is untested and probably contains lots of bugs, it is
primarily meant to sketch and flesh out the new structure of the HTML
generation.

The new HTML generation is simpler to understand, better to maintain,
more flexible and likely more efficient as well at the cost of using a
bit more memory.