Add printf function; split global config
There is now a function to encapsulate the recurring call to
print(string.format()) in printf().

The global CONF table is now split to create a second runtime config
table that can not be modified by config files such as rc.lua. This new
table is STATE and currently contains srcdir, storing the path passed
to lenie via command argument. This split is mainly for code hygene.